Dispatch event in case job fails and setting tries / retry behavior on jobs
