+++ src/Events/RequestCertificateFailed.php
@@ -2,6 +2,7 @@
 
 namespace Daanra\LaravelLetsEncrypt\Events;
 
+use Daanra\LaravelLetsEncrypt\Interfaces\NotifiableEvent;
 use Illuminate\Broadcasting\Channel;
 use Illuminate\Broadcasting\InteractsWithSockets;
 use Illuminate\Broadcasting\PresenceChannel;
@@ -10,7 +11,7 @@ use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
 use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Queue\SerializesModels;
 
-class RequestCertificateFailed
+class RequestCertificateFailed implements NotifiableEvent
 {
     use Dispatchable, InteractsWithSockets, SerializesModels;
