+++ src/Interfaces/NotifiableEvent.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Daanra\LaravelLetsEncrypt\Interfaces;
+
+interface NotifiableEvent
+{
+    public function __construct($data);
+}
