+++ src/Jobs/RegisterAccount.php
@@ -4,6 +4,7 @@ namespace Daanra\LaravelLetsEncrypt\Jobs;
 
 use Daanra\LaravelLetsEncrypt\Facades\LetsEncrypt;
 use Daanra\LaravelLetsEncrypt\Events\RegisterAccountFailed;
+use Daanra\LaravelLetsEncrypt\Traits\JobTrait;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
@@ -12,32 +13,10 @@ use Illuminate\Queue\SerializesModels;
 
 class RegisterAccount implements ShouldQueue
 {
-    use Dispatchable, Queueable, InteractsWithQueue, SerializesModels;
+    use Dispatchable, Queueable, InteractsWithQueue, SerializesModels, JobTrait;
 
     /** @var string|null */
     protected $email;
-
-    /**
-     * The number of times the job may be attempted.
-     *
-     * @var int
-     */
-    public $tries;
-
-    /**
-     * The number of seconds to wait before retrying the job.
-     *
-     * @var int
-     */
-    public $retryAfter;
-
-    /**
-     * The list of seconds to wait before retrying the job.
-     *
-     * @var array
-     */
-    public $retryList;
-
 
     public function __construct(string $email = null, int $tries = null, int $retryAfter = null, $retryList = [])
     {
@@ -51,16 +30,6 @@ class RegisterAccount implements ShouldQueue
     {
         $client = LetsEncrypt::createClient();
         $client->registerAccount(null, $this->email);
-    }
-
-    /**
-     * Calculate the number of seconds to wait before retrying the job.
-     *
-     * @return int
-     */
-    public function retryAfter()
-    {
-        return (!empty($this->retryList)) ? $this->retryList[$this->attempts() - 1] : 0;
     }
 
     /**
